practice/chapter8/code/requirements.txt
# 8장 실습: RAG의 기본과 프로덕션 체크리스트
# 설치: pip install -r requirements.txt

langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
openai>=1.50.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
